Rename i18n gulp tasks & fix code format
