+++ tools/tasks/seed/i18n.build.ts
@@ -2,6 +2,5 @@ import * as runSequence from 'run-sequence';
 import Config from '../../config';
 
 export = () => {
-    Config.BUILD_TYPE = 'prod';
-    runSequence('clean.tmp','copy.prod','build.html_css');
+  runSequence('clean.tmp','copy.prod','build.html_css');
 };
